Treat money / cash as an int (#31531)
This is one case where a truly dedicated player could potentially gather
the $20M+ required to break money formatting.  But we don't think it's a
serious concern.

+++ src/output.h
@@ -803,7 +803,7 @@ std::string format_volume( const units::volume &volume );
 std::string format_volume( const units::volume &volume, int width, bool *out_truncated,
                            double *out_value );
 
-inline const std::string format_money( signed long cents )
+inline const std::string format_money( int cents )
 {
     return string_format( _( "$%.2f" ), cents / 100.0 );
 }
